CafeDetailModel common 모듈로 이동
