+++ common/src/main/kotlin/us/wedemy/eggeum/android/common/model/CafeDetailModel.kt
@@ -5,7 +5,7 @@
  * Please see full license: https://github.com/Wedemy/eggeum-android/blob/main/LICENSE
  */
 
-package us.wedemy.eggeum.android.main.model
+package us.wedemy.eggeum.android.common.model
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
