+++ common/src/main/kotlin/us/wedemy/eggeum/android/common/model/ImageModel.kt
@@ -5,13 +5,12 @@
  * Please see full license: https://github.com/Wedemy/eggeum-android/blob/main/LICENSE
  */
 
-package us.wedemy.eggeum.android.main.model
+package us.wedemy.eggeum.android.common.model
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
 
 @Parcelize
-data class ProductModel(
-  val name: String,
-  val price: Int,
+data class ImageModel(
+  val files: List<FileModel> = emptyList(),
 ) : Parcelable
