+++ common/src/main/kotlin/us/wedemy/eggeum/android/common/model/MenuModel.kt
@@ -5,13 +5,12 @@
  * Please see full license: https://github.com/Wedemy/eggeum-android/blob/main/LICENSE
  */
 
-package us.wedemy.eggeum.android.updatecafe.model
+package us.wedemy.eggeum.android.common.model
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
 
 @Parcelize
-data class FileModel(
-  val uploadFileId: Int?,
-  val url: String?,
+data class MenuModel(
+  val products: List<ProductModel> = emptyList(),
 ) : Parcelable
