+++ common/src/main/kotlin/us/wedemy/eggeum/android/common/model/ProductModel.kt
@@ -5,12 +5,13 @@
  * Please see full license: https://github.com/Wedemy/eggeum-android/blob/main/LICENSE
  */
 
-package us.wedemy.eggeum.android.updatecafe.model
+package us.wedemy.eggeum.android.common.model
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
 
 @Parcelize
-data class MenuModel(
-  val products: List<ProductModel> = emptyList(),
+data class ProductModel(
+  val name: String,
+  val price: Int,
 ) : Parcelable
