--- main/src/main/kotlin/us/wedemy/eggeum/android/main/model/MenuModel.kt
@@ -1,16 +1,0 @@
-/*
- * Designed and developed by Wedemy 2023.
- *
- * Licensed under the MIT.
- * Please see full license: https://github.com/Wedemy/eggeum-android/blob/main/LICENSE
- */
-
-package us.wedemy.eggeum.android.main.model
-
-import android.os.Parcelable
-import kotlinx.parcelize.Parcelize
-
-@Parcelize
-data class MenuModel(
-  val products: List<ProductModel> = emptyList(),
-) : Parcelable
