+++ main/src/main/kotlin/us/wedemy/eggeum/android/main/ui/MainActivity.kt
@@ -21,7 +21,7 @@ import javax.inject.Inject
 import us.wedemy.eggeum.android.common.base.BaseActivity
 import us.wedemy.eggeum.android.main.R
 import us.wedemy.eggeum.android.main.databinding.ActivityMainBinding
-import us.wedemy.eggeum.android.main.model.CafeDetailModel
+import us.wedemy.eggeum.android.common.model.CafeDetailModel
 import us.wedemy.eggeum.android.navigator.LoginNavigator
 import us.wedemy.eggeum.android.navigator.UpdateCafeNavigator
