--- update-cafe/src/main/kotlin/us/wedemy/eggeum/android/updatecafe/model/ProductModel.kt
@@ -1,17 +1,0 @@
-/*
- * Designed and developed by Wedemy 2023.
- *
- * Licensed under the MIT.
- * Please see full license: https://github.com/Wedemy/eggeum-android/blob/main/LICENSE
- */
-
-package us.wedemy.eggeum.android.updatecafe.model
-
-import android.os.Parcelable
-import kotlinx.parcelize.Parcelize
-
-@Parcelize
-data class ProductModel(
-  val name: String,
-  val price: Int,
-) : Parcelable
